+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.3.1] - 2026-09-04
+
+### Added
+
+- Added Slot Highlight Behind Item tweak to draw most of the slot hover highlight underneath the item instead of on top
+  of it.
+    - Backported from 1.21.2, using the same `slot_highlight_back` and `slot_highlight_front`
+      sprites under the vanilla paths, so existing resource packs should work as expected.
+
 ## [3.3.0] - 2026-09-04
 
 ### Added